furquan-spring-minet-app/src/main/java/com/demo/minet/rest/TransactionRestController.java
package com.demo.minet.rest;

import com.demo.minet.entity.Transaction;
import com.demo.minet.service.TransactionService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
public class TransactionRestController {
    @Autowired
    private TransactionService transactionService;

    @GetMapping("/transactions")
    public List<Transaction> getAllTransactions() {
        return transactionService.getAllTransactions();
    }

    @GetMapping("/transactions/{id}")
    public Optional<Transaction> getTransactionById(@PathVariable int id) {
        return transactionService.getTransactionById(id);
    }

    @PostMapping("/transactions")
    public Transaction saveTransaction(@RequestBody Transaction transaction) {
        if (transaction != null) {
           transaction = transactionService.saveTransaction(transaction);
        }
        return transaction;
    }
}
